+++ Assets/Scripts/GameScripts/Classes/StrategicObjectives/Town.cs
@@ -7,17 +7,17 @@ public class Town : StratObj {
 		SupplyLevel = 50;
 	}
 
-    public new void MoneyToDefences(int Money) {
-        this.DefenceLevel += Money / getUpgradeDefenceCost();
+    public new void MoneyToDefences(float Money) {
+        this.DefenceLevel += (int)(Money / getUpgradeDefenceCost());
     }
-    public new void MoneyToSupply(int Money) {
-        this.SupplyLevel += Money / getUpgradeSupplyCost();
+    public new void MoneyToSupply(float Money) {
+        this.SupplyLevel += (int)(Money / getUpgradeSupplyCost());
     }
-    public new int getUpgradeDefenceCost() {
-        return (int)((1 / 500) * Mathf.Pow(DefenceLevel, 2.0f));
+    public new float getUpgradeDefenceCost() {
+        return ((1 / 500) * Mathf.Pow(DefenceLevel, 2.0f));
     }
-    public new int getUpgradeSupplyCost() {
-        return (int)((1 / 500) * Mathf.Pow(SupplyLevel, 1.9f));
+    public new float getUpgradeSupplyCost() {
+        return ((1 / 500) * Mathf.Pow(SupplyLevel, 1.9f));
     }
 
 	public override string ToString (){
